rename file to foreign_chain_configuration
