+++ crates/contract/tests/sandbox/mod.rs
@@ -1,5 +1,5 @@
 pub mod contract_configuration;
-pub mod foreign_chain_policy;
+pub mod foreign_chain_configuration;
 pub mod foreign_chain_request;
 pub mod participants_gas;
 pub mod sign;
